fixed crash when server would try to return to an inexistent lobby

+++ src/GameTS/Application/Services/LobbyService.ts
@@ -74,6 +74,7 @@ class LobbyService {
 
     getLobbyInfoForClient(lobbyID: number, userID: number) {
         const lobby = lobbyService.getLobbyByID(lobbyID);
+        if (!lobby) {throw Error (`Lobby ${lobbyID} not found when it should have been`)}
         const matchConfigs = matchService.getMatchClientConfigsByUserID(userID);
         if (matchConfigs) {
             const match = matchService.getMatchByUserID(userID);
@@ -96,6 +97,7 @@ class LobbyService {
 
     addUser(lobbyID: number, userID: number, username: string, spriteID: number, rating: number) {
         const lobby = lobbyRepository.getByID(lobbyID);
+        if (!lobby) {throw Error (`Lobby ${lobbyID} not found when it should have been`)}
         const user = {
             id: userID,
             username: username,
@@ -112,6 +114,7 @@ class LobbyService {
 
     removeUser(lobbyID: number, userID: number) {
         const lobby = lobbyRepository.getByID(lobbyID);
+        if (!lobby) {throw Error (`Lobby ${lobbyID} not found when it should have been`)}
         lobby.users = lobby.users.filter(user => user.id !== userID);
         socketService.broadcastToLobby(lobbyID, "removeLobbyUser", { userID: userID })
         //Automatically close lobby if nobody is in there and no match is active
@@ -122,6 +125,7 @@ class LobbyService {
 
     updateSettings(lobbyID: number, senderID: number, newSettings: MatchSettingsT) {
         const lobby = lobbyRepository.getByID(lobbyID);
+        if (!lobby) {throw Error (`Lobby ${lobbyID} not found when it should have been`)}
         if (lobby.hostID !== senderID) {
             console.log("Somehow a non host managed to send a settings change request! It will be ignored.");
             return;
@@ -223,6 +227,7 @@ class LobbyService {
 
     addTournamentPlayer(lobbyID: number, userID: number) {
         const lobby = lobbyRepository.getByID(lobbyID);
+        if (!lobby) {throw Error (`Lobby ${lobbyID} not found when it should have been`)}
         const participantsAmount = lobby.users.filter(user => user.player != null).length;
 
         if (participantsAmount < tournamentService.MAX_PARTICIPANTS) {
@@ -254,6 +259,7 @@ class LobbyService {
 
     start(lobbyID: number, senderID: number) {
         const lobby = lobbyRepository.getByID(lobbyID);
+        if (!lobby) {throw Error (`Lobby ${lobbyID} not found when it should have been`)}
         if (lobby.hostID != senderID) {
             console.log(`Somehow userID ${senderID} is not host and tried to start the game! The host id is ${lobby.hostID}`)
             return ;
@@ -282,6 +288,10 @@ class LobbyService {
 
     returnToLobby(lobbyID: number) {
         const lobby = lobbyRepository.getByID(lobbyID);
+        if (!lobby) {
+            console.log("The server tried to return to the lobby but everybody left... :( Ignoring");
+            return;
+        }
         if (lobby.users.length === 0) {
             lobbyRepository.remove(lobbyID);
         } else {
@@ -297,6 +307,7 @@ class LobbyService {
 
     private _getLobbyUserByID(lobbyID: number, userID: number) {
         const lobby = lobbyRepository.getByID(lobbyID);
+        if (!lobby) {throw Error (`Lobby ${lobbyID} not found when it should have been`)}
         const user = lobby.users.find(user => user.id === userID);
         if (!user) {throw Error("User could not be found in lobby!")};
         return user;
